refactor(api-types): Extract agentSkillSchema into agent-skill.schema.ts (no-changelog)
Pure move out of the DTO layer, mirroring agent-task.schema.ts, so schema
files (inline-agent-config.schema.ts next) can import the skill body schema
without reaching into dto.ts. Package exports are unchanged.
