+++ packages/@n8n/api-types/src/agents/index.ts
@@ -1,6 +1,7 @@
 export * from './agent-files.constants';
 export * from './agent-integration.schema';
 export * from './agent-json-config.schema';
+export * from './agent-skill.schema';
 export * from './inline-agent-config.schema';
 export * from './sanitize-agent-json-config';
 export * from './agent-task.schema';
